GP-3800: Get TerminalService working on Windows, too.

+++ Ghidra/Features/Base/src/main/java/ghidra/app/plugin/core/terminal/TerminalListener.java
@@ -28,7 +28,7 @@ public interface TerminalListener {
 	 * @param cols the number of columns
 	 * @param rows the number of rows
 	 */
-	default void resized(int cols, int rows) {
+	default void resized(short cols, short rows) {
 	}
 
 	/**
